rename: 에러 핸들링하는 인터페이스 이름 변경
