--- app/src/main/java/woowacourse/movie/base/BasePresenter.kt
@@ -1,3 +1,0 @@
-package woowacourse.movie.base
-
-interface BasePresenter
