--- app/src/main/java/woowacourse/movie/base/BaseView.kt
@@ -1,5 +1,0 @@
-package woowacourse.movie.base
-
-interface BaseView {
-    fun handleError()
-}
