+++ app/src/main/java/woowacourse/movie/feature/complete/MovieReservationCompleteContract.kt
@@ -1,12 +1,11 @@
 package woowacourse.movie.feature.complete
 
-import woowacourse.movie.base.BasePresenter
-import woowacourse.movie.base.BaseView
 import woowacourse.movie.model.Ticket
 import woowacourse.movie.model.data.dto.MovieContent
+import woowacourse.movie.utils.BasePresenter
 
 interface MovieReservationCompleteContract {
-    interface View : BaseView {
+    interface View {
         fun setUpMovieContentUi(movieContent: MovieContent)
 
         fun setUpTicketUi(ticket: Ticket)
