+++ app/src/main/java/woowacourse/movie/feature/reservation/MovieReservationContract.kt
@@ -1,11 +1,10 @@
 package woowacourse.movie.feature.reservation
 
-import woowacourse.movie.base.BasePresenter
-import woowacourse.movie.base.BaseView
 import woowacourse.movie.model.data.dto.MovieContent
+import woowacourse.movie.utils.BasePresenter
 
 interface MovieReservationContract {
-    interface View : BaseView {
+    interface View {
         fun setUpMovieContentUi(movieContent: MovieContent)
 
         fun updateReservationCountUi(reservationCount: Int)
