+++ app/src/main/java/woowacourse/movie/utils/BaseActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.base
+package woowacourse.movie.utils
 
 import androidx.appcompat.app.AppCompatActivity
