+++ app/src/main/java/woowacourse/movie/utils/BasePresenter.kt
@@ -1,0 +1,3 @@
+package woowacourse.movie.utils
+
+interface BasePresenter
