+++ app/src/main/java/woowacourse/movie/utils/ErrorListener.kt
@@ -1,0 +1,5 @@
+package woowacourse.movie.utils
+
+interface ErrorListener {
+    fun handleError()
+}
